ADAPTIVE FOOTER
A very small script that adds the shortContent class to the footer if it finds that the body is shorter than the window. In this case, css is applied to the shortContent class so that it sticks to the bottom of the page. It will act as usual on larger pages where the user can scroll.