Team Page created
This commit includes a page to visualize a team, including a list of players, the tournaments it's participated in, and a list of the most recent matches. A team.css file has been created to add some specific css to the small carousel that shows tournaments. There's no interactivity yet, though.

Some extra changes: the favicon has been modified to better reflect the website, and the page now has a custom title on some pages.