Styles Changes
Reformatted the RegisterAccount htmls, which were divided into two parts and converted it in only one html. Also added a functionality that returns to RegisterMatchSelectMatch html
In the tournaments sheet the format given to the imageviews of the players has changed. Also a carousel has been added.
A lot of styles changes