Stylesheet unification plus a bunch of page specific changes
The stylesheet has been unified with a previous version that was discarded due to an error reversing a commit. On top of this, lots of individual changes have been made, with a singular purpose: adapting the page structure to all possible screen sizes. A bunch of small changes have also been made for aesthetic purposes.