Tournament Page made more responsive
tournamentSheet.html is now even more responsible, even on the smallest screens. In them, the group phase will have smaller text so that everything fits nicely, and the cards will all be on a single column. The final phase, due to its special structure, will remain the same, but the user will have to swipe horizontally to see it in full to preserve its structure. We included a small message that informs the user of the ability to do this.