Adaptive recent matches
The team page now has an adaptive list of matches. Previously, it always showed 10 elements, and the mustache was not adapted for list elements at all. Now, with a new query we get a list of matches played by the team, and we display them accordingly. The next step is to only send the first 10 matches to the page and reload with ajax with more after that.