Pie chart tweak
There's a special variable in the pie chart called --over50, which has to be added to the inline style of a region of the chart only if this region has a value higher than 50. We added some mustache templates that check this and only add this clause in the appropiate cases. In TeamController we check this, and if that's the case, we mark over50lost or over50won with true.