Autocomplete searchboxes on main page
The tournament and team searchboxes on the team page will now show a list of suggestions as the user types. These suggestions depend on what the user writes, and they are dynamically added to a list in a script via a mustache template that retrieves a list of all teams and tournaments from a query to the database. We used the jquery ui autocomplete plugin to add the local functionality, and our usual jdbc calls and controllers to fill the list of teams and tournaments on the page dynamically.