Team and tournament search fully functional
The team and tournament search boxes in the main page are now fully functional. As the user types, a list of suggestions will pop up. If the users submit a name of a team or tournament that exists, the search abr will lead them directly to the appropiate page. If there's no team or tournament with that exact name, the user will be directed to a results page that offers a list of teams or tournaments with similar names, with working links to them. If there's no similar names, a message will notify the user too.